+++ src/sensors/interface.hpp
@@ -61,7 +61,7 @@ class TemperatureInterface {
    * @brief returns int representation of temperature
    * @return int temperature degrees C
    */
-  virtual int getData() = 0;
+  virtual uint8_t getData() = 0;
 };
 
 class PressureInterface {
